+++ Framework/DataHandling/src/LoadBBY.cpp
@@ -376,12 +376,12 @@ void LoadBBY::exec() {
   // currently beam monitor counts are not available, instead number of frames
   // times period is used
   logManager.addProperty(
-      "bm_counts", eventCounter.numFrames() * period /
+      "bm_counts", static_cast<double>(eventCounter.numFrames()) * period /
                        1.0e6); // static_cast<double>(instrumentInfo.bm_counts)
 
   // currently
   Kernel::time_duration duration = boost::posix_time::microseconds(
-      static_cast<boost::int64_t>(eventCounter.numFrames() * period));
+      static_cast<boost::int64_t>(static_cast<double>(eventCounter.numFrames()) * period));
 
   Kernel::DateAndTime start_time("2000-01-01T00:00:00");
   Kernel::DateAndTime end_time(start_time + duration);
